+++ lib/legion/extensions/pushover.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'legion/extensions/pushover/version'
+require 'legion/extensions/pushover/client'
 
 module Legion
   module Extensions